refactor: ExhibitionModalおよびFloorModalの非アニメーション要素をExhibitionDetailに移行
