+++ src/components/shared/ExhibitionDetail/ExhibitionDetail.tsx
@@ -1,10 +1,26 @@
 "use client";
 
-import React, { useEffect, useState } from 'react';
-import { Typography, Chip, Box, Tabs, Tab, Rating, TextField, Button } from '@mui/material';
+import React, { useEffect, useState, useRef } from 'react';
+import { 
+  Typography, 
+  Chip, 
+  Box, 
+  Tabs, 
+  Tab, 
+  Rating, 
+  TextField, 
+  Button,
+  DialogContent,
+  IconButton
+} from '@mui/material';
+import {
+  Close,
+  Favorite,
+} from '@mui/icons-material';
 import { ZoomIn, ImageNotSupported, Star, StarBorder } from '@mui/icons-material';
-import styles from '@/src/components/pages/map/ExhibitionModal/ExhibitionModal.module.css';
+import { motion, AnimatePresence } from 'framer-motion';
 import { supabase } from '@/src/lib/supabase/client';
+import styles from './ExhibitionDetail.module.css';
 
 interface Review {
   id: number;
@@ -32,23 +48,17 @@ interface ExhibitionItem {
   tags?: string[];
 }
 
-const StarRating: React.FC<{ rating: number }> = ({ rating }) => {
-  return (
-    <div className={styles.starRating}>
-      {[1, 2, 3, 4, 5].map((star) => (
-        <span key={star} className={styles.star}>
-          {star <= rating ? (
-            <Star fontSize="small" style={{ color: '#ffd700' }} />
-          ) : (
-            <StarBorder fontSize="small" style={{ color: '#ddd' }} />
-          )}
-        </span>
-      ))}
-    </div>
-  );
+type ExhibitionDetailProps = {
+  open: boolean;
+  onClose: () => void;
+  exhibition: ExhibitionItem | null;
 };
 
-const ExhibitionDetail: React.FC<{ exhibition: ExhibitionItem | null }> = ({ exhibition }) => {
+const ExhibitionDetail: React.FC<ExhibitionDetailProps> = ({
+  open,
+  onClose,
+  exhibition
+}) => {
   const [imageLoaded, setImageLoaded] = useState(false);
   const [imageError, setImageError] = useState(false);
   const [isImageZoomed, setIsImageZoomed] = useState(false);
@@ -56,6 +66,7 @@ const ExhibitionDetail: React.FC<{ exhibition: ExhibitionItem | null }> = ({ exh
   const [rating, setRating] = useState<number | null>(0);
   const [comment, setComment] = useState('');
   const [reviews, setReviews] = useState<Review[]>([]);
+  const closeButtonRef = useRef<HTMLButtonElement>(null);
 
   useEffect(() => {
     if (!exhibition?.id) return;
@@ -117,147 +128,195 @@ const ExhibitionDetail: React.FC<{ exhibition: ExhibitionItem | null }> = ({ exh
     }
   };
 
+  // Focus management for accessibility
+  useEffect(() => {
+    if (open && closeButtonRef.current) {
+      closeButtonRef.current.focus();
+    }
+  }, [open]);
+
+  useEffect(() => {
+    const handleKeyDown = (event: KeyboardEvent) => {
+      if (event.key === 'Escape' && open) {
+        onClose();
+      }
+    };
+
+    if (open) {
+      document.addEventListener('keydown', handleKeyDown);
+      document.body.style.overflow = 'hidden';
+    }
+
+    return () => {
+      document.removeEventListener('keydown', handleKeyDown);
+      document.body.style.overflow = 'unset';
+    };
+  }, [open, onClose]);
+
   if (!exhibition) return null;
 
   return (
-    <div className={styles.dialogContent}>
-      {/* Image */}
-      <div className={styles.imageSection}>
-        {exhibition.imageUrl ? (
-          <div
-            className={`${styles.imageContainer} ${isImageZoomed ? styles.imageZoomed : ''}`}
-            onClick={() => { if (imageLoaded && !imageError) setIsImageZoomed(!isImageZoomed); }}
-            role="button"
-            tabIndex={0}
-            onKeyDown={(e) => { if (e.key === 'Enter' && imageLoaded && !imageError) setIsImageZoomed(!isImageZoomed); }}
-            aria-label={exhibition.name + ' の画像'}
-          >
-            {!imageError ? (
-              <img
-                src={exhibition.imageUrl}
-                alt={exhibition.name}
-                className={styles.exhibitionImage}
-                onLoad={() => { setImageLoaded(true); setImageError(false); }}
-                onError={() => { setImageError(true); setImageLoaded(false); }}
-              />
+    <div className={styles.exhibitionDetailContainer}>
+      <IconButton
+        ref={closeButtonRef}
+        onClick={onClose}
+        className={styles.closeButton}
+        aria-label="モーダルを閉じる"
+        sx={{
+          position: 'absolute',
+          top: 16,
+          right: 16,
+          zIndex: 10,
+          backgroundColor: 'rgba(70, 48, 48, 0.5)',
+          color: 'white',
+          '&:hover': {
+            backgroundColor: 'rgba(0, 0, 0, 0.7)',
+          },
+        }}
+      >
+        <Close />
+      </IconButton>
+      <DialogContent className={styles.dialogContent}>
+        <div className={styles.dialogContent}>
+          {/* Image */}
+          <div className={styles.imageSection}>
+            {exhibition.imageUrl ? (
+              <div
+                className={`${styles.imageContainer} ${isImageZoomed ? styles.imageZoomed : ''}`}
+                onClick={() => { if (imageLoaded && !imageError) setIsImageZoomed(!isImageZoomed); }}
+                role="button"
+                tabIndex={0}
+                onKeyDown={(e) => { if (e.key === 'Enter' && imageLoaded && !imageError) setIsImageZoomed(!isImageZoomed); }}
+                aria-label={exhibition.name + ' の画像'}
+              >
+                {!imageError ? (
+                  <img
+                    src={exhibition.imageUrl}
+                    alt={exhibition.name}
+                    className={styles.exhibitionImage}
+                    onLoad={() => { setImageLoaded(true); setImageError(false); }}
+                    onError={() => { setImageError(true); setImageLoaded(false); }}
+                  />
+                ) : (
+                  <div className={styles.imagePlaceholder}>
+                    <ImageNotSupported className={styles.placeholderIcon} />
+                    <div className={styles.placeholderText}>画像を読み込めませんでした</div>
+                  </div>
+                )}
+
+                <div className={styles.imageOverlay}>
+                  <ZoomIn className={styles.zoomIcon} />
+                  <span className={styles.zoomText}>拡大する</span>
+                </div>
+              </div>
             ) : (
               <div className={styles.imagePlaceholder}>
                 <ImageNotSupported className={styles.placeholderIcon} />
-                <div className={styles.placeholderText}>画像を読み込めませんでした</div>
+                <div className={styles.placeholderText}>画像がありません</div>
               </div>
             )}
-
-            <div className={styles.imageOverlay}>
-              <ZoomIn className={styles.zoomIcon} />
-              <span className={styles.zoomText}>拡大する</span>
-            </div>
           </div>
-        ) : (
-          <div className={styles.imagePlaceholder}>
-            <ImageNotSupported className={styles.placeholderIcon} />
-            <div className={styles.placeholderText}>画像がありません</div>
-          </div>
-        )}
-      </div>
 
-      {/* Title */}
-      <div className={styles.titleSection}>
-        <Typography variant="h4" component="h1" className={styles.exhibitionTitle}>
-          {exhibition.name}
-        </Typography>
-        <div className={styles.metaInfo}>
-          <div className={styles.metaItem}>
-            <span>{exhibition.location}</span>
-          </div>
-          <div className={styles.metaItem}>
-            <span>{exhibition.schedule}</span>
-          </div>
-          {exhibition.capacity && (
-            <div className={styles.metaItem}>
-              <span>定員 {exhibition.capacity}名</span>
-            </div>
-          )}
-        </div>
-      </div>
-
-      {/* Tags */}
-      <div className={styles.tagsSection}>
-        {(exhibition.tags || []).map((tag, index) => {
-          const tagStyle = getTagColor(tag);
-          return (
-            <Chip
-              key={tag}
-              label={tag}
-              className={styles.tag}
-              sx={{ backgroundColor: tagStyle.bg, color: tagStyle.color, fontWeight: 600, fontSize: '0.8rem' }}
-            />
-          );
-        })}
-      </div>
-
-      {/* Reviews / Tabs */}
-      <div className={styles.reviewsSection}>
-        <Box className={styles.tabsContainer}>
-          <Tabs value={tabValue} onChange={(_, newValue) => setTabValue(newValue)} className={styles.tabs}>
-            <Tab label="説明" />
-            <Tab label="レビュー" />
-          </Tabs>
-        </Box>
-
-        <Box className={styles.tabContent}>
-          {tabValue === 0 && (
-            <div>
-              <div className={styles.descriptionSection}>
-                <Typography variant="h6" className={styles.descriptionTitle}>展示説明</Typography>
-                <div className={styles.descriptionContent}>
-                  <Typography variant="body1" className={styles.descriptionText}>
-                    {exhibition.detailedDescription || exhibition.description}
-                  </Typography>
-                  {exhibition.organizer && (
-                    <div className={styles.organizerInfo}>
-                      <Typography variant="body2" className={styles.organizerLabel}>主催者:</Typography>
-                      <Typography variant="body2" className={styles.organizerName}>{exhibition.organizer}</Typography>
-                    </div>
-                  )}
+          {/* Title */}
+          <div className={styles.titleSection}>
+            <Typography variant="h4" component="h1" className={styles.exhibitionTitle}>
+              {exhibition.name}
+            </Typography>
+            <div className={styles.metaInfo}>
+              <div className={styles.metaItem}>
+                <span>{exhibition.location}</span>
+              </div>
+              <div className={styles.metaItem}>
+                <span>{exhibition.schedule}</span>
+              </div>
+              {exhibition.capacity && (
+                <div className={styles.metaItem}>
+                  <span>定員 {exhibition.capacity}名</span>
                 </div>
-              </div>
+              )}
+            </div>
+          </div>
 
-              <div className={styles.reviewsList}>
-                {reviews.length > 0 ? (
-                  <p>レビューが投稿されております</p>
-                ) : (
-                  <div className={styles.noReviews}>
-                    <Typography variant="body1" color="textSecondary">まだレビューがありません。最初のレビューを投稿してみませんか？</Typography>
+          {/* Tags */}
+          <div className={styles.tagsSection}>
+            {(exhibition.tags || []).map((tag, index) => {
+              const tagStyle = getTagColor(tag);
+              return (
+                <Chip
+                  key={tag}
+                  label={tag}
+                  className={styles.tag}
+                  sx={{ backgroundColor: tagStyle.bg, color: tagStyle.color, fontWeight: 600, fontSize: '0.8rem' }}
+                />
+              );
+            })}
+          </div>
+
+          {/* Reviews / Tabs */}
+          <div className={styles.reviewsSection}>
+            <Box className={styles.tabsContainer}>
+              <Tabs value={tabValue} onChange={(_, newValue) => setTabValue(newValue)} className={styles.tabs}>
+                <Tab label="説明" />
+                <Tab label="レビュー" />
+              </Tabs>
+            </Box>
+
+            <Box className={styles.tabContent}>
+              {tabValue === 0 && (
+                <div>
+                  <div className={styles.descriptionSection}>
+                    <Typography variant="h6" className={styles.descriptionTitle}>展示説明</Typography>
+                    <div className={styles.descriptionContent}>
+                      <Typography variant="body1" className={styles.descriptionText}>
+                        {exhibition.detailedDescription || exhibition.description}
+                      </Typography>
+                      {exhibition.organizer && (
+                        <div className={styles.organizerInfo}>
+                          <Typography variant="body2" className={styles.organizerLabel}>主催者:</Typography>
+                          <Typography variant="body2" className={styles.organizerName}>{exhibition.organizer}</Typography>
+                        </div>
+                      )}
+                    </div>
                   </div>
-                )}
-              </div>
-            </div>
-          )}
 
-          {tabValue === 1 && (
-            <div className={styles.reviewForm}>
-              <Box className={styles.ratingSection}>
-                <Typography variant="h6" className={styles.ratingLabel}>評価 <span className={styles.required}>*</span></Typography>
-                <Rating size="large" value={rating} onChange={(_, newValue) => setRating(newValue)} className={styles.rating} />
-              </Box>
+                  <div className={styles.reviewsList}>
+                    {reviews.length > 0 ? (
+                      <p>レビューが投稿されております</p>
+                    ) : (
+                      <div className={styles.noReviews}>
+                        <Typography variant="body1" color="textSecondary">まだレビューがありません。最初のレビューを投稿してみませんか？</Typography>
+                      </div>
+                    )}
+                  </div>
+                </div>
+              )}
 
-              <Box className={styles.commentSection}>
-                <Typography variant="h6" className={styles.commentLabel}>コメント</Typography>
-                <TextField fullWidth multiline rows={4} variant="outlined" placeholder="体験の感想をお聞かせください..." value={comment} onChange={(e) => setComment(e.target.value)} className={styles.textField} inputProps={{ maxLength: 500 }} helperText={`${comment.length}/500文字`} />
-              </Box>
+              {tabValue === 1 && (
+                <div className={styles.reviewForm}>
+                  <Box className={styles.ratingSection}>
+                    <Typography variant="h6" className={styles.ratingLabel}>評価 <span className={styles.required}>*</span></Typography>
+                    <Rating size="large" value={rating} onChange={(_, newValue) => setRating(newValue)} className={styles.rating} />
+                  </Box>
 
-              <Button onClick={handleSubmitReview} variant="contained" disabled={!rating} className={styles.submitButton} sx={{ alignSelf: 'center' }}>レビューを投稿</Button>
+                  <Box className={styles.commentSection}>
+                    <Typography variant="h6" className={styles.commentLabel}>コメント</Typography>
+                    <TextField fullWidth multiline rows={4} variant="outlined" placeholder="体験の感想をお聞かせください..." value={comment} onChange={(e) => setComment(e.target.value)} className={styles.textField} inputProps={{ maxLength: 500 }} helperText={`${comment.length}/500文字`} />
+                  </Box>
 
-              <h2>レビュー一覧</h2>
-              <ul>
-                {reviews.map((review) => (
-                  <li key={review.id}>{review.comment} (評価:{review.rating})</li>
-                ))}
-              </ul>
-            </div>
-          )}
-        </Box>
-      </div>
+                  <Button onClick={handleSubmitReview} variant="contained" disabled={!rating} className={styles.submitButton} sx={{ alignSelf: 'center' }}>レビューを投稿</Button>
+
+                  <h2>レビュー一覧</h2>
+                  <ul>
+                    {reviews.map((review) => (
+                      <li key={review.id}>{review.comment} (評価:{review.rating})</li>
+                    ))}
+                  </ul>
+                </div>
+              )}
+            </Box>
+          </div>
+        </div>
+      </DialogContent>
     </div>
   );
 };
